% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelperSpatial.R
\name{getRowsAndCols}
\alias{getRowsAndCols}
\title{Internal: Determine rows and columns for arbitrary number of plots}
\usage{
getRowsAndCols(numplots = NA, nrow = NA, ncol = NA)
}
\arguments{
\item{numplots}{A numeric indicating the number of plots available for plot.}

\item{nrow}{A numeric indicating number of rows in the plot.}

\item{ncol}{A numeric indicating number of columns in the plot.}
}
\value{
A 2 value vector wih first value denoting row and second value as column.
}
\description{
Internal: Determine rows and columns for figures from arbitrary number of plots
}

